Chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-community==0.2.16
langchain-text-splitters==0.2.4
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
pypdf==4.3.1
docx2txt==0.8
faiss-cpu==1.8.0
numpy==1.26.4